Hopefully makes Django testing work nicely with a local Informix DB
There seems to be no way get PyODBC/unixODBC/Informix to *really* close
a connection completely. As such when Django testing closes the DB
connection, opens an nodb_connection and issues a `DROP DATABASE` it
fails as there are open connections. Even when a dirty `lseep` is issued
to give Informix time to clean up.

I've tried manually disabling connection pooling at the pyODBC level,
setting CPTime to zero, and `CONN_MAX_AGE` to zero with no success. As
well as a myriad of other unpleasant approaches.  Nothing worked.

So, this PR wraps the destroy DB in a try/except and just output a
simple message when dropping the DB fails. It also overrides the create
test db process to automatically destroy and existing test database (if
it exist) rather than prompt the user. The `keepdb` option still does
the expected thing.

So from the developers point of view Django tests do exactly what they
would do normally, whilst in point of fact doing pretty much the
opposite.

I also updated the `DatabaseSchemaEditor.execute` function to ignore
index creation errors where the index already exists. This is bacause
Informix automatically adds indexes to foreign keys but Django is not
expecting this so try to do it effectively a second time. This was
leading to some unncessary and confusing looking warning messages in the
test output.

Als, it turns out that the IBM informix dev database Docker images does the
creation of a database at start up by passing in some environmental
variables. It just isn't mentioned in their readme.

So I have added a docker-compose file with the relevant variables set
and updated the test project settings to match