Removing shortcut in SQL compiler
The SQL compiler class for Informix contained a shortcut if the high and lower marks for limit/offset where the same. It would skip any furthe computation if they were the same. In doing so it also skipped a whole bunch of additional logic that set stuff up for pagination.